Intraday/requirements.txt

---

### **`requirements.txt`**
```txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
requests
pyyaml
joblib
